Add parsing for constants
A note on the addressing mode for "constant", it's needed since we at
parsing cannot know the size of `LDA zpage` and `LDA abs_addr` and thus
not determine if its a zeropage address or an absolute address. This has
to be done after the symbol resolution when we've resolved the
constants.