Remove the assembly code dependency on the compiler
I realized that the compiler module was tightly coupled to the Lexer,
Parser *and* the AST while it really just needs to be coupled to the AST
module. So this patch rewrites the compile module to only accept an AST
as input which makes the code base a little bit more decoupled.

I'm thinking that it would be good to do the same analysis on the other
modules as well. The end-to-end tests should rather be done in the
integration tests.