asm: Add support for .org (origin) directive
This will replace the program start concept that up til this point has
been the way to set where in memory the program is stored or where it
will be compiled to.

Now, as shown in examples/loop.asm,`.org $8000` specifies that the
program will be placed at address `0x8000`. The resulting binary from
the assembler will be padded with zeroes if no code is generated up til
the address specified by the origin directive.